include構成を変更

+++ src/Utility/utility.cpp
@@ -2,8 +2,8 @@
 #include "utility.hpp"
 
 
-float toRadians(float deg) {
-  return float(deg * M_PI) / 180.0f;
+double toRadians(float deg) {
+  return (deg * M_PI) / 180.0f;
 }
 
 Vec3f nearPosOnLine(const Vec3f& p,
@@ -16,10 +16,8 @@ Vec3f nearPosOnLine(const Vec3f& p,
 
   AB.normalize();
 
-  // A‚©‚çNearPos‚Ü‚Å‚Ì‹——£
-  double distAX = AB.dot(AP);
+  float distAX = AB.dot(AP);
 
-  // NearPos
   Vec3f ret;
   ret = vA + (AB * distAX);
